move,remove zerodisk
