+++ openvmm/openvmm_entry/src/lib.rs
@@ -27,8 +27,8 @@ use cli_args::EndpointConfigCli;
 use cli_args::NicConfigCli;
 use cli_args::SerialConfigCli;
 use cli_args::UefiConsoleModeCli;
+use disk_backend_resources::layer::RamDiskLayerHandle;
 use disk_backend_resources::DiskLayerHandle;
-use disk_backend_resources::RamDiskLayerHandle;
 use floppy_resources::FloppyDiskConfig;
 use framebuffer::FramebufferAccess;
 use framebuffer::FRAMEBUFFER_SIZE;
