+++ openvmm/openvmm_resources/src/lib.rs
@@ -42,7 +42,7 @@ vm_resource::register_static_resolvers! {
     net_dio::resolver::DioResolver,
 
     // Disks
-    disk_backend::layered::resolver::LayeredDiskResolver,
+    disk_layered::resolver::LayeredDiskResolver,
     #[cfg(feature = "disk_crypt")]
     disk_crypt::resolver::DiskCryptResolver,
     disk_ramdisk::resolver::RamDiskResolver,
