+++ petri/src/vm/construct.rs
@@ -18,9 +18,9 @@ use crate::BOOT_NVME_NSID;
 use crate::SCSI_INSTANCE;
 use crate::SIZE_1_GB;
 use anyhow::Context;
+use disk_backend_resources::layer::RamDiskLayerHandle;
 use disk_backend_resources::DiskLayerHandle;
 use disk_backend_resources::LayeredDiskHandle;
-use disk_backend_resources::RamDiskLayerHandle;
 use framebuffer::Framebuffer;
 use framebuffer::FramebufferAccess;
 use framebuffer::FRAMEBUFFER_SIZE;
