+++ vm/devices/storage/disk_backend_resources/src/lib.rs
@@ -6,6 +6,8 @@
 #![warn(missing_docs)]
 #![forbid(unsafe_code)]
 
+pub mod layer;
+
 use mesh::MeshPayload;
 use vm_resource::kind::DiskHandleKind;
 use vm_resource::kind::DiskLayerHandleKind;
@@ -15,21 +17,6 @@ use vm_resource::ResourceId;
 
 // Define config types here so that you don't have to pull in the individual
 // crates just to describe the configuration.
-
-/// RAM disk layer handle.
-///
-/// FUTURE: allocate shared memory here so that the disk can be migrated between
-/// processes.
-#[derive(MeshPayload)]
-pub struct RamDiskLayerHandle {
-    /// The size of the layer. If `None`, the layer will be the same size as the
-    /// lower disk.
-    pub len: Option<u64>,
-}
-
-impl ResourceId<DiskLayerHandleKind> for RamDiskLayerHandle {
-    const ID: &'static str = "ram";
-}
 
 /// File-backed disk handle.
 #[derive(MeshPayload)]
