+++ vm/devices/storage/disk_layered/Cargo.toml
@@ -2,11 +2,13 @@
 # Licensed under the MIT License.
 
 [package]
-name = "disk_backend"
+name = "disk_layered"
 edition = "2021"
 rust-version.workspace = true
 
 [dependencies]
+disk_backend.workspace = true
+disk_backend_resources.workspace = true
 scsi_buffers.workspace = true
 
 guestmem.workspace = true
@@ -15,7 +17,6 @@ inspect = { workspace = true, features = ["std"] }
 
 async-trait.workspace = true
 futures.workspace = true
-stackfuture.workspace = true
 thiserror.workspace = true
 
 [lints]
