+++ vm/devices/storage/disk_layered/src/bitmap.rs
