+++ vm/devices/storage/disk_layered/src/resolve.rs
