+++ vm/devices/storage/disk_layered/src/resolver.rs
@@ -9,10 +9,10 @@ use super::DiskLayer;
 use super::InvalidLayeredDisk;
 use super::LayerConfiguration;
 use super::LayeredDisk;
-use crate::resolve::ResolveDiskParameters;
-use crate::resolve::ResolvedDisk;
-use crate::InvalidDisk;
 use async_trait::async_trait;
+use disk_backend::resolve::ResolveDiskParameters;
+use disk_backend::resolve::ResolvedDisk;
+use disk_backend::InvalidDisk;
 use disk_backend_resources::DiskLayerHandle;
 use disk_backend_resources::LayeredDiskHandle;
 use futures::future::TryJoinAll;
