+++ vm/devices/storage/disk_ramdisk/Cargo.toml
@@ -9,6 +9,7 @@ rust-version.workspace = true
 [dependencies]
 disk_backend.workspace = true
 disk_backend_resources.workspace = true
+disk_layered.workspace = true
 guestmem.workspace = true
 scsi_buffers.workspace = true
 vm_resource.workspace = true
