+++ vmm_tests/vmm_tests/tests/tests/x86_64/openhcl_linux_direct.rs
@@ -5,7 +5,7 @@
 
 use anyhow::Context;
 use disk_backend_resources::LayeredDiskHandle;
-use disk_backend_resources::RamDiskLayerHandle;
+use disk_backend_resources::layer::RamDiskLayerHandle;
 use gdma_resources::GdmaDeviceHandle;
 use gdma_resources::VportDefinition;
 use guid::Guid;
